pyproject.toml
[project]
name = "doc-scan-processor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dash-ag-grid>=32.3.2",
    "dash-iconify>=0.1.2",
    "dash-mantine-components>=2.4.0",
    "easyocr>=1.7.2",
    "notebook>=7.4.7",
    "opencv-python>=4.12.0.88",
    "pandas>=2.3.3",
    "pillow>=12.0.0",
    "plotly>=6.4.0",
    "scikit-optimize>=0.10.2",
    "scipy>=1.16.3",
    "tqdm>=4.67.1",
]
